requirements.txt
# --- Core runtime dependencies (RAG PDF + Gradio) ---
gradio
langchain
langchain-community
langchain-google-genai
langchain-qdrant
qdrant-client
pypdf
python-dotenv
sentence-transformers

# SDKs de Google (usa uno u otro según tu código)
google-generativeai     # para langchain_google_genai (más común)
google-genai            # si usas: from google import genai

# --- Opcionales (descomenta si los usas) ---
# fastapi
# uvicorn[standard]
# rank-bm25
# ragas
# pydantic-settings
# tiktoken
